Move run_basic_passes_on to module
