+++ qirlib/src/evaluation/jit.rs
@@ -6,7 +6,7 @@ use crate::evaluation::{
     intrinsics::{reset_max_qubit_id, reset_static_result_cache, set_measure_stream},
     runtime::Simulator,
 };
-use crate::{module, passes::run_basic_passes_on};
+use crate::module;
 use bitvec::prelude::BitVec;
 use inkwell::{
     attributes::AttributeLoc,
@@ -58,7 +58,7 @@ pub fn run_module(
         return Err("Target doesn't have a target machine.".to_owned());
     }
 
-    run_basic_passes_on(module);
+    module::run_basic_passes_on(module);
     let entry_point = choose_entry_point(module_functions(module), entry_point)?;
 
     // load the symbols for the current process (empty/null string)
