+++ qirlib/src/module.rs
@@ -2,8 +2,14 @@
 // Licensed under the MIT License.
 
 use inkwell::{
-    attributes::AttributeLoc, builder::Builder, context::Context, memory_buffer::MemoryBuffer,
-    module::Module, values::FunctionValue,
+    attributes::AttributeLoc,
+    builder::Builder,
+    context::Context,
+    memory_buffer::MemoryBuffer,
+    module::Module,
+    passes::{PassManager, PassManagerBuilder},
+    values::FunctionValue,
+    OptimizationLevel,
 };
 use std::{ffi::OsStr, path::Path};
 
@@ -92,6 +98,17 @@ pub fn build_entry_point(module: &Module, builder: &Builder) {
     let entry_point = create_entry_point(module);
     let entry = context.append_basic_block(entry_point, "entry");
     builder.position_at_end(entry);
+}
+
+// This method returns true if any of the passes modified the function or module and false otherwise.
+pub fn run_basic_passes_on(module: &Module) -> bool {
+    let pass_manager_builder = PassManagerBuilder::create();
+    pass_manager_builder.set_optimization_level(OptimizationLevel::None);
+    let fpm = PassManager::create(());
+    fpm.add_global_dce_pass();
+    fpm.add_strip_dead_prototypes_pass();
+    pass_manager_builder.populate_module_pass_manager(&fpm);
+    fpm.run_on(module)
 }
 
 fn create_entry_point<'ctx>(module: &Module<'ctx>) -> FunctionValue<'ctx> {
